requirements.txt
asgiref==3.5.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2022.5.18.1
cffi==1.15.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.0.12
click==8.1.3
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
defusedxml==0.7.1
Django==4.0.4
django-cors-headers==3.12.0
django-decouple==2.1
django-phonenumber-field==6.1.0
django-qr-code==3.0.0
djangorestframework==3.13.1
ecdsa==0.17.0
ffmpeg-python==0.2.0
filelock==3.12.3
Flask==2.1.2
future==0.18.3
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.4
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.40.1
MarkupSafe==2.1.1
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.0
openai-whisper==20230314
optional-django==0.3.0
Pillow==9.1.1
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycparser==2.21
pydub==0.25.1
PyJWT==2.4.0
pyOpenSSL==23.2.0
python-dotenv==0.20.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.1
react==4.3.0
redis==5.0.0
regex==2023.8.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
segno==1.4.1
service-identity==23.1.0
shortuuid==1.0.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
sympy==1.12
tiktoken==0.3.1
tomlkit==0.10.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==1.26.9
Werkzeug==2.1.2
whisper==1.1.10
zipp==3.8.0
zope.interface==6.0
